Fix item info in manual comparison file
To save runtime when merging the recommendations (that are just item ids,
so not interpretable by humans) with item info from the "items.csv" file,
we initially subsetted the item info to the items for which recommendations
should be made. However, we actually need item info for the items that are
recommended. Also, the merge is very fast even if the item dataset is not
subsetted at all. Thus, we just remove this statement.