Adapt task 2 pre-processing to another dataset
Support another auction dataset which has larger domains of prices.
Fortunately, format is nearly the same as in the existing dataset, so we
can extract the pre-processing routine and call it for both datasets.

Still some changes required in pre-processing of new dataset:
- string matching to extract allocation slightly adapted
- permutation numbering adapted to count for multiple final allocations
per permutation (as the might be multiple potential winners for each
product)
- two new ids (existing two ids still suffice to identify, but the two new
ones help for group-by)
- fix (make consistent to small dataset) values of winner column
- fix (make consistent to small dataset) capacities for rows where a
winner is found